+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For semantic convention package changes, see the [semconv CHANGELOG](packages/se
 ### :boom: Breaking Change
 
 * feat(sdk-metrics): Add support for aggregation cardinality limit with a default limit of 2000. This limit can be customized via views [#5182](https://github.com/open-telemetry/opentelemetry-js/pull/5128)
+* fix(sdk-metrics): InMemoryMetricExporter clear metrics after shutdown to align with other exporters [#5131](https://github.com/open-telemetry/opentelemetry-js/issues/5131) @paper2
 
 ### :rocket: (Enhancement)